WS24WebService/requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
pymongo==4.6.1
python-dotenv==1.0.0
WTForms==3.1.1
dnspython==2.4.2
gunicorn==21.2.0
